Relocate mshv::heki_service into the litebox_service_heki crate
Move the VSM/HEKI service modules (vsm, state, mem_integrity) out of
litebox_platform_lvbs::mshv::heki_service into a standalone litebox_service_heki
crate. The service code itself is unchanged except crate-relative module paths
and the runner's import path.

Beyond mechanical relocation, the crate split enforces the Vtl0Mediation
capability boundary introduced in the previous commit. As its own crate, the
service has no access to litebox_platform_lvbs internals (the protected-frame
registry, VTL0 mappers, control-register locking) and no crypto dependencies of
its own to reach them. It can therefore act on the platform ONLY through the
Vtl0Mediation capability it is handed at construction plus the public
litebox_common_lvbs wire types -- turning the trait-based capability into a
compiler-enforced wall rather than a convention.
