+++ .github/workflows/ci.yml
@@ -288,6 +288,10 @@ jobs:
           #
           # - `litebox_runner_snp` is `no_std` but requires custom target to build
           #
+          # - `litebox_service_heki` is `no_std` but with crypto dependencies (sha2,
+          #   rsa, ...) whose soft-float/asm code fails to compile for
+          #   `x86_64-unknown-none`.
+          #
           # - `dev_tests` is meant to only be used for tests, and thus can
           #   safely use std.
           #
@@ -307,6 +311,7 @@ jobs:
             -not -path './litebox_syscall_rewriter/Cargo.toml' \
             -not -path './litebox_packager/Cargo.toml' \
             -not -path './litebox_runner_snp/Cargo.toml' \
+            -not -path './litebox_service_heki/Cargo.toml' \
             -not -path './dev_tests/Cargo.toml' \
             -not -path './dev_bench/Cargo.toml' \
             -print0 | \
