+++ Cargo.toml
@@ -15,6 +15,7 @@ members = [
     "litebox_runner_lvbs",
     "litebox_runner_optee_on_linux_userland",
     "litebox_shim_linux",
+    "litebox_service_heki",
     "litebox_syscall_rewriter",
     "litebox_packager",
     "litebox_runner_snp",
@@ -39,6 +40,7 @@ default-members = [
     "litebox_runner_linux_on_windows_userland",
     "litebox_shim_linux",
     "litebox_shim_optee",
+    "litebox_service_heki",
     "litebox_syscall_rewriter",
     "litebox_packager",
     "litebox_util_log",
