+++ litebox_platform_lvbs/Cargo.toml
@@ -10,7 +10,6 @@ litebox = { path = "../litebox/", version = "0.1.0" }
 litebox_common_linux = { path = "../litebox_common_linux/", version = "0.1.0" }
 litebox_common_lvbs = { path = "../litebox_common_lvbs/", version = "0.1.0" }
 litebox_util_log = { version = "0.1.0", path = "../litebox_util_log" }
-log = { version = "0.4", default-features = false }
 spin = { version = "0.10.0", default-features = false, features = [
   "spin_mutex",
   "once",
@@ -23,14 +22,7 @@ num_enum = { version = "0.7.3", default-features = false }
 once_cell = { version = "1.20.2", default-features = false, features = ["alloc", "race"] }
 modular-bitfield = { version = "0.12.0", default-features = false }
 hashbrown = "0.15.2"
-elf = { version = "0.8.0", default-features = false }
-cms = { version = "0.2.3", default-features = false, features = ["alloc"] }
-rsa = { version = "0.9.10", default-features = false }
 sha2 = { version = "0.10.9", default-features = false, features = ["oid"] }
-x509-cert = { version = "0.2.5", default-features = false }
-const-oid = { version = "0.9.6", default-features = false, features = ["db"] }
-authenticode = { version = "0.4.3", default-features = false, features = ["object"] }
-object = { version = "0.36.7", default-features = false, features = ["pe"] }
 digest = { version = "0.10.7", default-features = false }
 aligned-vec = { version = "0.6.4", default-features = false }
 raw-cpuid = "11.6.0"
