--- litebox_platform_lvbs/src/mshv/heki_service/mod.rs
@@ -1,8 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-pub mod mem_integrity;
-pub mod state;
-pub mod vsm;
-pub use state::HekiState;
-pub use vsm::Vsm;
