+++ litebox_service_heki/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "litebox_service_heki"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+litebox = { path = "../litebox/", version = "0.1.0" }
+litebox_common_lvbs = { path = "../litebox_common_lvbs/", version = "0.1.0" }
+litebox_common_linux = { path = "../litebox_common_linux/", version = "0.1.0" }
+zerocopy = { version = "0.8", default-features = false, features = ["derive"] }
+x86_64 = { version = "0.15.2", default-features = false, features = ["instructions"] }
+log = { version = "0.4", default-features = false }
+spin = { version = "0.10.0", default-features = false, features = [
+  "spin_mutex",
+  "once",
+  "rwlock",
+] }
+hashbrown = "0.15.2"
+rangemap = { version = "1.5.1", features = ["const_fn"] }
+thiserror = { version = "2.0.6", default-features = false }
+once_cell = { version = "1.20.2", default-features = false, features = ["alloc", "race"] }
+elf = { version = "0.8.0", default-features = false }
+cms = { version = "0.2.3", default-features = false, features = ["alloc"] }
+rsa = { version = "0.9.10", default-features = false }
+sha2 = { version = "0.10.9", default-features = false, features = ["oid"] }
+x509-cert = { version = "0.2.5", default-features = false }
+const-oid = { version = "0.9.6", default-features = false, features = ["db"] }
+authenticode = { version = "0.4.3", default-features = false, features = ["object"] }
+object = { version = "0.36.7", default-features = false, features = ["pe"] }
+digest = { version = "0.10.7", default-features = false }
+
+[lints]
+workspace = true
